feat:Modified migration files

+++ database/migrations/2021_10_20_073932_create_company_groups_table.php
@@ -14,16 +14,10 @@ class CreateCompanyGroupsTable extends Migration
     public function up()
     {
         Schema::create('company_groups', function (Blueprint $table) {
-            $table
-                ->bigIncrements('id');
-            $table
-                ->text('name')
-                ->comment('系列名');
-            $table
-                ->softDeletes()
-                ->comment('削除フラグ');
-            $table
-                ->timestamps();
+            $table->integer('id');
+            $table->text('name')->comment('系列名');
+            $table->softDeletes()->comment('削除フラグ');
+            $table->timestamps();
         });
     }
